Have KLayout run def2stream script directly
